Rename PyQtDWMG to PyDWMG
